requirements.txt
streamlit
selenium
beautifulsoup4
groq
python-dotenv
requests
